Refactor IndoorNavigation components to use specific style files for improved organization and maintainability
