+++ ConcordiaMaps/styles.js
@@ -239,197 +239,5 @@ const styles = StyleSheet.create({
   },
 });
 
-const buildingSelectorStyles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: "#fff",
-  },
-  contentContainer: {
-    flex: 1,
-    padding: 16,
-  },
-  title: {
-    fontSize: 28,
-    fontWeight: "bold",
-    textAlign: "center",
-    marginVertical: 20,
-    color: "#912338",
-  },
-  buildingsContainer: {
-    flex: 1,
-    gap: 20,
-  },
-  buildingCard: {
-    backgroundColor: "#f5f5f5",
-    borderRadius: 15,
-    overflow: "hidden",
-    elevation: 4,
-    shadowColor: "#000",
-    shadowOffset: { width: 0, height: 2 },
-    shadowOpacity: 0.25,
-    shadowRadius: 3.84,
-  },
-  buildingContent: {
-    padding: 20,
-  },
-  textContainer: {
-    gap: 8,
-  },
-  buildingName: {
-    fontSize: 24,
-    fontWeight: "bold",
-    color: "#333",
-  },
-  buildingCode: {
-    fontSize: 18,
-    fontWeight: "600",
-    color: "#912338",
-  },
-  buildingDescription: {
-    fontSize: 16,
-    color: "#666",
-    marginTop: 4,
-  },
-  buildingAddress: {
-    fontSize: 14,
-    color: "#888",
-    marginTop: 4,
-  },
-  availableTag: {
-    position: "absolute",
-    top: 10,
-    right: 10,
-    backgroundColor: "#912338",
-    color: "#fff",
-    padding: 8,
-    borderRadius: 12,
-    fontSize: 12,
-    fontWeight: "bold",
-  },
-});
 
-const floorSelectorStyles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: "#fff",
-  },
-  title: {
-    fontSize: 24,
-    fontWeight: "bold",
-    textAlign: "center",
-    marginTop: 16,
-    color: "#912338",
-  },
-  subtitle: {
-    fontSize: 18,
-    textAlign: "center",
-    marginBottom: 16,
-    color: "#666",
-  },
-  carousel: {
-    flex: 1,
-  },
-  floorCard: {
-    margin: 10,
-    padding: 20,
-    backgroundColor: "#f5f5f5",
-    borderRadius: 12,
-    elevation: 3,
-    shadowColor: "#000",
-    shadowOffset: { width: 0, height: 2 },
-    shadowOpacity: 0.25,
-    shadowRadius: 3.84,
-  },
-  activeFloor: {
-    backgroundColor: "#e6f3ff",
-    borderColor: "#912338",
-    borderWidth: 2,
-  },
-  floorNumber: {
-    fontSize: 28,
-    fontWeight: "bold",
-    marginBottom: 8,
-    color: "#333",
-  },
-  floorName: {
-    fontSize: 20,
-    fontWeight: "600",
-    marginBottom: 8,
-    color: "#444",
-  },
-  floorDescription: {
-    fontSize: 16,
-    color: "#666",
-  },
-});
-
-const tunnelNavigationStyles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: "#fff",
-    padding: 16,
-  },
-  title: {
-    fontSize: 28,
-    fontWeight: "bold",
-    textAlign: "center",
-    marginVertical: 20,
-    color: "#912338",
-  },
-  subtitle: {
-    fontSize: 16,
-    color: "#666",
-    textAlign: "center",
-    marginBottom: 24,
-  },
-  buildingsGrid: {
-    flex: 1,
-    gap: 20,
-  },
-  buildingCard: {
-    backgroundColor: "#f5f5f5",
-    borderRadius: 15,
-    overflow: "hidden",
-    elevation: 4,
-    shadowColor: "#000",
-    shadowOffset: { width: 0, height: 2 },
-    shadowOpacity: 0.25,
-    shadowRadius: 3.84,
-  },
-  buildingContent: {
-    padding: 20,
-  },
-  textContainer: {
-    gap: 8,
-  },
-  buildingName: {
-    fontSize: 24,
-    fontWeight: "bold",
-    color: "#333",
-  },
-  buildingCode: {
-    fontSize: 18,
-    fontWeight: "600",
-    color: "#912338",
-  },
-  buildingDescription: {
-    fontSize: 16,
-    color: "#666",
-    marginTop: 4,
-  },
-  buildingAddress: {
-    fontSize: 14,
-    color: "#888",
-    marginTop: 4,
-  },
-  scrollContainer: {
-    flex: 1,
-  },
-});
-
-export default {
-  styles,
-  buildingSelectorStyles,
-  floorSelectorStyles,
-  tunnelNavigationStyles,
-};
+export default styles
